Added, different tabs in maintence, componentatized inventory section
