--- app/login/Login.module.css
@@ -1,8 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  width: 100vw;
-  background: linear-gradient(-26.5deg, #fafafa 50%, #eccb34 50%);
-}
