+++ app/property-management/CleaningSection.jsx
@@ -37,6 +37,7 @@ const CleaningSection = ({
   const [errorDialogOpen, setErrorDialogOpen] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
   const [selectedDate, setSelectedDate] = useState(dayjs());
+  const [confirmSubmitDialogOpen, setConfirmSubmitDialogOpen] = useState(false);
 
   const handleDateChange = useCallback(
     (newDate) => {
@@ -126,7 +127,7 @@ const CleaningSection = ({
 
       if (isFileAttached && fileAttached) {
         const fileBase64 = await convertFileToBase64(fileAttached);
-        const uploadRes = await fetchWithAuth("/api/upload/cleaning", {
+        const uploadRes = await fetchWithAuth("/api/upload/invoice", {
           method: "POST",
           body: JSON.stringify({
             propertyName: selectedPropertyName,
@@ -405,7 +406,14 @@ const CleaningSection = ({
             },
             minWidth: "200px",
           }}
-          onClick={handleCleaningSubmit}
+          onClick={() => {
+            const newErrors = validateInputs();
+            if (Object.keys(newErrors).length > 0) {
+              setErrors(newErrors);
+              return;
+            }
+            setConfirmSubmitDialogOpen(true);
+          }}
         >
           Submit Cleaning
         </Button>
@@ -510,6 +518,65 @@ const CleaningSection = ({
             className="bg-primary text-dark hover:bg-primary/80 transition-colors"
           >
             Close
+          </Button>
+        </DialogActions>
+      </Dialog>
+      <Dialog
+        open={confirmSubmitDialogOpen}
+        onClose={() => setConfirmSubmitDialogOpen(false)}
+        PaperProps={{
+          sx: {
+            backgroundColor: "#fafafa",
+            color: "#333333",
+            borderRadius: "12px",
+            border: "1px solid rgba(236, 203, 52, 0.2)",
+          },
+        }}
+      >
+        <DialogTitle sx={{ color: "#333333" }}>Confirm Submission</DialogTitle>
+        <DialogContent>
+          <DialogContentText sx={{ color: "#333333" }}>
+            Please confirm you want to submit the following cleaning invoice:
+          </DialogContentText>
+          <div className="mt-4 space-y-2 text-dark">
+            <p>
+              <strong>Property:</strong> {selectedPropertyName}
+            </p>
+            <p>
+              <strong>Company:</strong> {selectedCompany}
+            </p>
+            <p>
+              <strong>Cost:</strong> ${Number(cleaningCost).toFixed(2)}
+            </p>
+            <p>
+              <strong>Date:</strong>{" "}
+              {dayjs(selectedDate).format("MMMM D, YYYY")}
+            </p>
+            <p>
+              <strong>Description:</strong>{" "}
+              {cleaningDescription || "None provided"}
+            </p>
+            <p>
+              <strong>File:</strong>{" "}
+              {fileAttached ? fileAttached.name : "None attached"}
+            </p>
+          </div>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={() => setConfirmSubmitDialogOpen(false)}
+            className="text-dark hover:bg-primary/5 transition-colors"
+          >
+            Cancel
+          </Button>
+          <Button
+            onClick={() => {
+              setConfirmSubmitDialogOpen(false);
+              handleCleaningSubmit();
+            }}
+            className="bg-primary text-dark hover:bg-primary/80 transition-colors"
+          >
+            Confirm Submission
           </Button>
         </DialogActions>
       </Dialog>
