--- app/property-management/addPage.module.css
@@ -1,241 +1,0 @@
-.addPageContainer {
-  padding: 2rem;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-}
-
-.filterBarContainer {
-  width: 15vw;
-  height: auto;
-}
-
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-}
-
-.contentContainer {
-  display: flex;
-  flex: 1;
-  padding-bottom: 5rem;
-}
-
-.leftContainer {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  position: relative;
-}
-
-.productListContainer {
-  flex: 1;
-  padding: 2rem;
-  background: transparent;
-  border-radius: 8px;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-  overflow-y: auto;
-}
-
-.buttonContainer {
-  padding: 1rem 2rem;
-  background-color: transparent;
-  display: flex;
-  gap: 1rem;
-  margin-top: 1rem;
-  z-index: 2;
-  align-items: center;
-}
-
-.buttonContainer button {
-  min-width: 100px;
-  height: 40px;
-}
-
-.pdfSection {
-  display: inline-block;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  font-weight: bold;
-  margin-bottom: 0.5rem;
-  color: #eccb34;
-}
-
-.header span:first-child {
-  margin-left: 0.5rem;
-}
-
-.header span:nth-child(2) {
-  margin-right: 2rem;
-}
-
-.listContainer {
-  flex: 1;
-  overflow-y: auto;
-  padding-right: 10px;
-  margin-bottom: 1rem;
-  max-height: calc(100vh - 250px);
-  color: #fafafa;
-}
-
-.productRow {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.5rem;
-  border-radius: 4px;
-  background-color: #eccb34;
-  margin-bottom: 0.5rem;
-  min-height: 40px;
-}
-
-.productName {
-  font-size: 0.9rem;
-  color: #fafafa;
-}
-
-.amountInput {
-  width: 60px;
-  padding: 0.3rem;
-  font-size: 0.9rem;
-  text-align: center;
-  border: none;
-  border-radius: 10;
-  background-color: #fafafa;
-  color: #eccb34;
-}
-
-.maintenanceInput {
-  width: auto;
-  padding: 1rem;
-  font-size: 0.9rem;
-  text-align: center;
-  border: 1px solid #fafafa;
-  border-radius: 8px;
-  background-color: #eccb34;
-  color: #fafafa;
-}
-
-.maintenanceInput::placeholder {
-  color: #fafafa;
-  opacity: 1;
-}
-
-.maintenanceInput::-webkit-input-placeholder {
-  color: #fafafa;
-  opacity: 1;
-}
-
-.maintenanceInput::-moz-placeholder {
-  color: #fafafa;
-  opacity: 1;
-}
-
-.maintenanceInput:-ms-input-placeholder {
-  color: #fafafa;
-  opacity: 1;
-}
-
-.maintenanceInput:focus {
-  border: 1px solid #fafafa;
-  outline: none;
-}
-
-.rightContainer {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  position: relative;
-}
-
-.maintenanceContainer {
-  position: relative;
-  z-index: 2;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 2rem;
-}
-
-.rightHeader,
-.leftHeader {
-  position: relative;
-  z-index: 2;
-  padding: 2rem;
-  font-weight: bold;
-  font-size: 1.2rem;
-  color: #fafafa;
-  text-align: center;
-  background-color: #eccb34;
-  border-radius: 8px;
-  border: none;
-  margin-bottom: 2rem;
-}
-
-.rightContentContainer {
-  position: relative;
-  z-index: 2;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.categoryContainer,
-.dateContainer,
-.companyContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  gap: 7rem;
-}
-
-.categoryContainer span,
-.dateContainer span,
-.companyContainer span {
-  font-weight: bold;
-  color: #2b2b2b; /* Keep dark gray for labels if present */
-}
-
-/* Style for buttons to match OptionBar */
-.Button {
-  background-color: #eccb34 !important; /* Yellow background */
-  color: #fafafa !important; /* White text */
-  border: 1px solid #2b2b2b !important; /* Keep border for buttons */
-  border-radius: 8px !important; /* Match OptionBar's border radius */
-  padding: 0.5rem 1rem !important;
-  min-width: 100px !important;
-  height: 40px !important;
-}
-
-/* Ensure the DatePicker (if styled similarly) matches */
-.DatePicker {
-  border: 1px solid #2b2b2b; /* Keep border for DatePicker */
-  border-radius: 8px; /* Match OptionBar's border radius */
-  background-color: #eccb34; /* Yellow background */
-  color: #fafafa; /* White text */
-  padding: 0.5rem;
-}
-
-/* Optional: If you want to override any Material-UI default styles for consistency */
-.MuiInputBase-root {
-  background-color: #eccb34 !important;
-  color: #fafafa !important;
-  border: 1px solid #2b2b2b !important; /* Keep border for Material-UI inputs */
-  border-radius: 8px !important;
-}
-
-.MuiInputBase-input {
-  color: #fafafa !important;
-}
