+++ lib/services/cleaningService.js
@@ -9,7 +9,7 @@ class CleaningService {
       console.log("Inserting Cleaning data:", CleaningData);
 
       await client.query(
-        `INSERT INTO Cleaning (
+        `INSERT INTO cleanings (
          property_uid, company, cost, description, date
         ) VALUES ($1, $2, $3, $4, $5)`,
         [
@@ -36,17 +36,31 @@ class CleaningService {
     try {
       const result = await client.query(
         `SELECT 
+          id,
           company,
           cost,
           description,
           date,
           TO_CHAR(date, 'YYYY-MM') as month
-        FROM Cleaning 
+        FROM cleanings
         WHERE property_uid = $1 
         ORDER BY date DESC`,
         [propertyId]
       );
       return result.rows;
+    } finally {
+      client.release();
+    }
+  }
+
+  static async deleteCleaning(cleaningId) {
+    if (typeof cleaningId === "undefined") {
+      throw new Error("cleaningId is required");
+    }
+
+    const client = await pool.connect();
+    try {
+      await client.query("DELETE FROM cleanings WHERE id = $1", [cleaningId]);
     } finally {
       client.release();
     }
@@ -65,7 +79,7 @@ class CleaningService {
             'description', description,
             'date', date
           )) as Cleaning_items
-        FROM Cleaning 
+        FROM cleanings
         WHERE property_uid = $1 
         GROUP BY TO_CHAR(date, 'YYYY-MM')
         ORDER BY month DESC`,
