+++ lib/services/googleService.js
@@ -13,16 +13,21 @@ console.log("Redirect URI:", process.env.GOOGLE_REDIRECT_URI);
 
 const drive = google.drive({ version: "v3", auth: oauth2Client });
 const sheets = google.sheets({ version: "v4", auth: oauth2Client });
+const gmail = google.gmail({ version: "v1", auth: oauth2Client });
 
 const googleService = {
   drive,
   sheets,
+  gmail,
 
   getAuthUrl: () => {
     const scopes = [
       "https://www.googleapis.com/auth/drive",
       "https://www.googleapis.com/auth/drive.file",
       "https://www.googleapis.com/auth/spreadsheets",
+      "https://www.googleapis.com/auth/gmail.readonly",
+      "https://www.googleapis.com/auth/gmail.send",
+      "https://www.googleapis.com/auth/gmail.compose",
     ];
     return oauth2Client.generateAuthUrl({
       access_type: "offline",
@@ -48,7 +53,7 @@ const googleService = {
       access_token: process.env.GOOGLE_ACCESS_TOKEN,
       refresh_token: process.env.GOOGLE_REFRESH_TOKEN,
       scope:
-        "https://www.googleapis.com/auth/drive https://www.googleapis.com/auth/drive.file https://www.googleapis.com/auth/spreadsheets",
+        "https://www.googleapis.com/auth/drive https://www.googleapis.com/auth/drive.file https://www.googleapis.com/auth/spreadsheets https://www.googleapis.com/auth/gmail.readonly https://www.googleapis.com/auth/gmail.send https://www.googleapis.com/auth/gmail.compose",
       token_type: "Bearer",
       expiry_date: process.env.GOOGLE_TOKEN_EXPIRY
         ? parseInt(process.env.GOOGLE_TOKEN_EXPIRY)
@@ -70,7 +75,10 @@ const googleService = {
 
   findReceiptsFolder: async (propertyName, monthYear) => {
     const propertyFolderName = `${propertyName} - Receipts`;
-    const propertyQuery = `name='${propertyFolderName}' mimeType='application/vnd.google-apps.folder'`;
+    const propertyQuery = `name = '${propertyFolderName.replace(
+      /'/g,
+      "\\'"
+    )}' and mimeType = 'application/vnd.google-apps.folder'`;
     console.log("Searching Drive with query:", propertyQuery);
     const propertyResponse = await drive.files.list({
       q: propertyQuery,
@@ -96,7 +104,10 @@ const googleService = {
     }
 
     const monthYearFolderName = monthYear;
-    const monthYearQuery = `'${propertyFolderId}' in parents name='${monthYearFolderName}' mimeType='application/vnd.google-apps.folder'`;
+    const monthYearQuery = `'${propertyFolderId}' in parents and name = '${monthYearFolderName.replace(
+      /'/g,
+      "\\'"
+    )}' and mimeType = 'application/vnd.google-apps.folder'`;
     console.log("Searching Drive with query:", monthYearQuery);
     const monthYearResponse = await drive.files.list({
       q: monthYearQuery,
