+++ lib/services/maintenanceService.js
@@ -37,6 +37,7 @@ class MaintenanceService {
     try {
       const result = await client.query(
         `SELECT 
+          id,
           category,
           company,
           cost,
@@ -49,6 +50,20 @@ class MaintenanceService {
         [propertyId]
       );
       return result.rows;
+    } finally {
+      client.release();
+    }
+  }
+
+  static async deleteMaintenance(maintenanceId) {
+    if (typeof maintenanceId === "undefined") {
+      throw new Error("maintenanceId is required");
+    }
+    const client = await pool.connect();
+    try {
+      await client.query("DELETE FROM maintenance WHERE id = $1", [
+        maintenanceId,
+      ]);
     } finally {
       client.release();
     }
